tests for ChecklistTaskController
